flowCTI/sub_agents/planning_agent/agent.py
from google.adk.agents import LlmAgent
from google.adk.tools import google_search

# Declare agent with built-in tool
def make_planning_agent():
    return LlmAgent(
    name="planning_agent",
    model="gemini-1.5-pro",
    description="Planning agent who specialised in cyber threat hunting.",
    instruction=("""You are news hunting on cyber activities around the globe.
                   You always have in hand the lastest news of the field.
                   Your duty include designinng security achitecture, protecting your organisation from data breach, financial loss, intellectual property theft, compromises, etc.
                   Answer user questions with profession in mind, strategically in protection planning with calm tone.
                   Be concise whenever you provide user with lists or planning related tasks. Your message should have medium amount of words.
                   Make a list for readability.
                   """),
    tools=[google_search]
)