requirements.txt
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
google-adk[database]==1.3.0
python-dotenv==1.1.0
google-generativeai==0.8.5
taxii2-client==2.3.0
stix2==3.0.1
deprecated==1.2.14